% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogetic_tree.R
\name{build_binary_table}
\alias{build_binary_table}
\title{Convert A TraceQC object to a matrix}
\usage{
build_binary_table(TraceQC_input, count_cutoff = 5)
}
\arguments{
\item{TraceQC_input}{A TraceQC object}

\item{count_cutoff}{The barcode cutoff. The function only considers barcodes
whose raw counts are equal or greater than the cutoff.}
}
\value{
It returns a list that contains a character matrix, a sequence
information of each barcode, and a data frame that contains code of each
barcode sequence.
}
\description{
Convert A TraceQC object to a matrix
}
