% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tree.R
\name{build_character_table}
\alias{build_character_table}
\title{Convert A TraceQC object to a matrix}
\usage{
build_character_table(TraceQC_input)
}
\arguments{
\item{TraceQC_input}{A TraceQC object}
}
\value{
It returns a list that contains a character matrix and a sequence
information.
}
\description{
Convert A TraceQC object to a matrix
}
