% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular.R
\name{circular_chordgram}
\alias{circular_chordgram}
\title{Display a circos function with links for a given data frame.}
\usage{
circular_chordgram(df, title, traceQC_input)
}
\arguments{
\item{traceQC_input}{}
}
\value{

}
\description{
Display a circos function with links for a given data frame.
}
