% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular.R
\name{circular_histogram}
\alias{circular_histogram}
\title{Display a circos function with a histgoram for a given data frame.}
\usage{
circular_histogram(df, title, traceQC_input)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Display a circos function with a histgoram for a given data frame.
}
