% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_object.R
\name{create_input_object_with_alignment}
\alias{create_input_object_with_alignment}
\title{Creating an input object for TraceQC with an alignment result.}
\usage{
create_input_object_with_alignment(aligned_reads_file, ref_file, fastqc_file)
}
\arguments{
\item{aligned_reads_file}{A path to store alignment output file.}

\item{ref_file}{A path of a reference sequence file.}

\item{fastqc_file}{A path of a FASTQC file.}
}
\value{
A list with those four elements.
\itemize{
  \item `aligned_reads': an data frame that contains alignment information.
  \item `refseq': The reference sequence.
  \item `regions': Detailed information about the reference sequence.
  \item `qc': A list of tibbles containing the QC data.
}
}
\description{
Creating an input object for TraceQC with an alignment result.
}
