% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutation_pattern_diff.R
\name{create_obj_list}
\alias{create_obj_list}
\title{Create a list of TraceQC objects from a list}
\usage{
create_obj_list(samples, ref, ncores = 4)
}
\arguments{
\item{samples}{A list that contains FASTQ file path.}

\item{ref}{The path to the reference file.}

\item{ncores}{The number of cores for parallel processing.}
}
\value{
It returns a list of TraceQC object
}
\description{
Create a list of TraceQC objects from a list
}
\examples{
\dontrun{
samples <- list(
  "day00" = system.file("extdata", "test_data", "fastq",
                        "example_0d.fastq.gz", package="TraceQC"),

  "day02" = system.file("extdata", "test_data", "fastq",
                        "example_2d.fastq.gz", package="TraceQC"),

  "day14" = system.file("extdata", "test_data", "fastq",
                        "example_14d.fastq.gz", package="TraceQC")
)

ref <- system.file("extdata", "test_data", "ref",
                   "ref.txt", package="TraceQC")

obj_list <- create_obj_list(samples, ref)
}
}
