% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_obj}
\alias{example_obj}
\title{An example TraceQC object}
\format{
An object of class \code{list} of length 5.
}
\usage{
data(example_obj)
}
\description{
An example TraceQC object
}
\keyword{datasets}
