% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_obj_list}
\alias{example_obj_list}
\title{An example of a list of TraceQC objects}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(example_obj_list)
}
\description{
An example of a list of TraceQC objects
}
\keyword{datasets}
