% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fastqc.file}
\alias{fastqc.file}
\title{Running FASTQC for a file.}
\usage{
fastqc.file(input_file, threads = 4)
}
\arguments{
\item{input_file}{A FASTQ file path}

\item{threads}{The number of threads will be used in FASTQC}
}
\value{
It returns a path to the QC file.
}
\description{
Running FASTQC for a file.
}
\examples{
input_file <- system.file("extdata", "test_data",
                          "fastq", "example.fastq.gz", package="TraceQC")
fastqc.file(input_file)
}
