% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell.R
\name{filter_mutations}
\alias{filter_mutations}
\title{Filter mutations based on read count per UMI}
\usage{
filter_mutations(data, include_max = TRUE, freq_threshold)
}
\arguments{
\item{data}{A data frame.}

\item{freq_threshold}{threshold of mutation frequency.}
}
\value{
A filtered data frame.
}
\description{
Filter mutations based on read count per UMI
}
