% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell.R
\name{filter_mutations_per_cell}
\alias{filter_mutations_per_cell}
\title{Filter mutations based on read count per cell}
\usage{
filter_mutations_per_cell(data, freq_threshold)
}
\arguments{
\item{data}{A data frame.}

\item{freq_threshold}{threshold of mutation frequency.}
}
\value{
A filtered data frame.
}
\description{
Filter mutations based on read count per cell
}
