% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report.R
\name{generate_qc_report}
\alias{generate_qc_report}
\title{Create a HTML QC report}
\usage{
generate_qc_report(
  input_file = NULL,
  ref_file = NULL,
  fastqc_dir = NULL,
  output_path = NULL,
  ncores = 4,
  title = "TraceQC report",
  preview = FALSE
)
}
\arguments{
\item{input_file}{A path of a FASTQ file.}

\item{ref_file}{A path of a text file which contains construct information.}

\item{fastqc_dir}{A directory path to store FastQC files. It will be omiited
if it is set as NULL}

\item{output_path}{A file path to save the HTML report. It will be omiited
if it is set as NULL}

\item{ncores}{The number of cores for parallel processing.}

\item{title}{The title of the report.}

\item{preview}{It will open a preview to a web-browser if the argument is set to `TRUE'}
}
\value{
A TraceQC object for the input will be returned.
}
\description{
Create a HTML QC report
}
\examples{
\dontrun{
library(TraceQC)
obj <- generate_qc_report(
  input_file = system.file("extdata", "test_data",
                           "fastq", "example_14d.fastq.gz", package="TraceQC"),
  ref_file = system.file("extdata", "test_data",
                         "ref", "ref.txt", package="TraceQC"),
  preview = FALSE,
  title = "TraceQC report for example_14d.fastq.gz",
  ncores=1
  )
summary(obj)
}
}
