% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report.R
\name{generate_qc_report_from_obj}
\alias{generate_qc_report_from_obj}
\title{Create a HTML QC report from TraceQC object}
\usage{
generate_qc_report_from_obj(
  traceQC_obj_path = NULL,
  output_path = NULL,
  ncores = 1,
  title = "TraceQC report",
  preview = FALSE
)
}
\arguments{
\item{traceQC_obj_path}{Path of a traceQC object.}

\item{output_path}{A file path to save the HTML report. It will be omiited
if it is set as NULL}

\item{ncores}{The number of cores for parallel processing.}

\item{title}{The title of the report.}

\item{preview}{It will open a preview to a web-browser if the argument is set to `TRUE'}
}
\value{
A TraceQC object for the input will be returned.
}
\description{
Create a HTML QC report from TraceQC object
}
