% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_abspath}
\alias{get_abspath}
\title{Get absolute path of a file.}
\usage{
get_abspath(f)
}
\arguments{
\item{f}{A relative or absolute file path.}
}
\value{
It returns an absolute path for a file.
}
\description{
Get absolute path of a file.
}
