% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_qcpath}
\alias{get_qcpath}
\title{Function to locate where the Quality Control file is stored.}
\usage{
get_qcpath(input_file, qc_dir)
}
\arguments{
\item{input_file}{A path of FASTQ file.}

\item{qc_dir}{The directory path where Quality Control files are.}
}
\value{
A path to the corresponded zipped Quality Control file.
}
\description{
Function to locate where the Quality Control file is stored.
}
