% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceQC_single_cell.R
\name{get_read_count_per_UB}
\alias{get_read_count_per_UB}
\title{Get read count per UMI table}
\usage{
get_read_count_per_UB(aligned_reads)
}
\arguments{
\item{aligned_reads}{A data frame of aligned reads}
}
\value{
A data frame of UMI
}
\description{
Get read count per UMI table
}
