% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell.R
\name{get_UMI_count_per_CB}
\alias{get_UMI_count_per_CB}
\title{Get UMI count per cell.}
\usage{
get_UMI_count_per_CB(df)
}
\arguments{
\item{df}{aligned reads}
}
\value{
A data frame of Cells.
}
\description{
Get UMI count per cell.
}
