% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{install_external_packages}
\alias{install_external_packages}
\title{Installation necessary external software for TraceQC}
\usage{
install_external_packages()
}
\description{
Installation necessary external software for TraceQC
}
\examples{
\dontrun{
library(TraceQC)
install_external_packages()
}
}
