% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lorenz_curve}
\alias{lorenz_curve}
\title{Drawing Lorenz Curve}
\usage{
lorenz_curve(traceQC_input)
}
\arguments{
\item{traceQC_input}{A TraceQC object}
}
\value{
A ggplot2 object that shows Lorenz Curve
}
\description{
The Lorenz curve shows an inequality of barcode distribution of the sample.
}
\examples{
data(example_obj)
lorenz_curve(example_obj)

}
