% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mutation_type}
\alias{mutation_type}
\title{A pie chart that shows a summary of mutation types.}
\usage{
mutation_type(traceQC_input)
}
\arguments{
\item{traceQC_input}{A TraceQC object}
}
\value{
A ggplot2 object that shows the pie chart
}
\description{
A pie chart that shows a summary of mutation types.
}
\examples{
data(example_obj)
mutation_type(example_obj)

}
