% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{mutation_type_donut}
\alias{mutation_type_donut}
\title{A pie chart that shows a summary of mutation types.}
\usage{
mutation_type_donut(mutations)
}
\arguments{
\item{mutations}{A mutation dataframe}
}
\value{
A ggplot2 object that shows the pie chart
}
\description{
A pie chart that shows a summary of mutation types.
}
\examples{
mutation <- read_tsv(system.file("extdata/test_data","mutation.txt",package="TraceQC"))
mutation_type_donut(mutation)

}
