% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{num_mutation_histogram}
\alias{num_mutation_histogram}
\title{A barplot to show distribution of the number of mutations per barcode}
\usage{
num_mutation_histogram(mutations)
}
\arguments{
\item{mutations}{A mutation dataframe}
}
\value{
A ggplot2 object that shows the barplot
}
\description{
A barplot to show distribution of the number of mutations per barcode
}
\examples{
data(example_obj)
num_mutation_histogram(example_obj)

}
