% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot_alignment_permutation}
\alias{plot_alignment_permutation}
\title{Visualization of alignment permutation.}
\usage{
plot_alignment_permutation(alignment_permutation)
}
\arguments{
\item{alignment_permutation}{an data frame of permutation sequence, output of `sequence_permutation`}
}
\value{
it returns A ggplot2 object that shows the permutation.
}
\description{
Visualization of alignment permutation.
}
