% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot_construct}
\alias{plot_construct}
\title{Visualization of the construct (reference sequence) information.}
\usage{
plot_construct(ref, chr_per_row = 50, chr_size = 10)
}
\arguments{
\item{ref}{an reference object, output of `parse_ref_file`}
}
\value{
it returns A ggplot2 object that shows the construct information.
}
\description{
Visualization of the construct (reference sequence) information.
}
