% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_plots.R
\name{plot_deletion_hotspot}
\alias{plot_deletion_hotspot}
\title{Display a circos plot that shows overall deletion pattern across the barcodes.}
\usage{
plot_deletion_hotspot(traceQC_input, count_cutoff = 1)
}
\arguments{
\item{traceQC_input}{A TraceQC object}

\item{count_cutoff}{A cutoff to remove link whose log10-count are less than the value.}
}
\value{
It doesn't generate any specific output.
}
\description{
Display a circos plot that shows overall deletion pattern across the barcodes.
}
\examples{
data(example_obj)
plot_deletion_hotspot(example_obj)

}
