% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_plots.R
\name{plot_insertion_hotspot}
\alias{plot_insertion_hotspot}
\title{Display a circos plot that shows overall insertion pattern across the barcodes.}
\usage{
plot_insertion_hotspot(mutations, ref, use_log_count = TRUE, count_cutoff = 1)
}
\arguments{
\item{mutations}{A mutations dataframe.}

\item{ref}{A reference object.}

\item{count_cutoff}{A cutoff to remove link whose count (or log10-count) are less than the value (default: 1).}
}
\value{
It won't return any specific object.
}
\description{
Display a circos plot that shows overall insertion pattern across the barcodes.
}
\examples{
ref_file <- system.file("extdata/test_data/ref","ref_hgRNA_invitro.txt",package="TraceQC")
ref <- parse_ref_file(ref_file)
mutation <- read_tsv(system.file("extdata/test_data","mutation.txt",package="TraceQC"))
plot_insertion_hotspot(mutation,ref)

}
