% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutation_pattern_diff.R
\name{plot_mutation_pattern_lineplot}
\alias{plot_mutation_pattern_lineplot}
\title{Plot the mutation pattern of given samples using a line plot.}
\usage{
plot_mutation_pattern_lineplot(obj_list)
}
\arguments{
\item{obj_list}{A list created by `create_obj_list'.}
}
\value{
It returns a ggplot object that contains line plot.
}
\description{
Plot the mutation pattern of given samples using a line plot.
}
\examples{
data(example_obj_list)
plot_mutation_pattern_lineplot(example_obj_list)
}
