% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutation_pattern_diff.R
\name{plot_mutation_pattern_violinplot}
\alias{plot_mutation_pattern_violinplot}
\title{Plot the mutation pattern of given samples using a violin plot.}
\usage{
plot_mutation_pattern_violinplot(obj_list)
}
\arguments{
\item{obj_list}{A list created by `create_obj_list'.}
}
\value{
It returns a ggplot object that contains violin plot.
}
\description{
Plot the mutation pattern of given samples using a violin plot.
}
\examples{
data(example_obj_list)
plot_mutation_pattern_violinplot(example_obj_list)
}
