% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogetic_tree.R
\name{plot_phylogenetic_tree}
\alias{plot_phylogenetic_tree}
\title{Plot a phylogenetic tree of a tree object using ggtree}
\usage{
plot_phylogenetic_tree(tree_obj)
}
\arguments{
\item{tree_obj}{A variable which was created by `table_to_phyDat'}
}
\value{
It returns a ggplot2 object of a phylogenetic tree.
}
\description{
Plot a phylogenetic tree of a tree object using ggtree
}
\examples{
library(TraceQC)
library(magrittr)
library(cowplot)
library(ggplot2)

data("example_obj_list")
p1 <- example_obj_list$day00 \%>\%
build_binary_table \%>\%
table_to_phyDat \%>\%
plot_phylogenetic_tree
p2 <- example_obj_list$day02 \%>\%
build_binary_table \%>\%
table_to_phyDat \%>\%
plot_phylogenetic_tree
p3 <- example_obj_list$day14 \%>\%
build_binary_table \%>\%
table_to_phyDat \%>\%
plot_phylogenetic_tree

plot_grid(p1 + ggtitle("Day 0"),
          p2 + ggtitle("Day 2"),
          p3 + ggtitle("Day 14"),
          nrow=1)

}
