% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{seq_split}
\alias{seq_split}
\title{Split a string by a fixed length and}
\usage{
seq_split(s, len = 50)
}
\arguments{
\item{s}{The input string}

\item{len}{The fixed length}
}
\value{
A string splited by the `len' then joined by `<br/>'
}
\description{
Split a string by a fixed length and
}
