% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_events.R
\name{seq_to_character}
\alias{seq_to_character}
\title{Identifying mutation events.}
\usage{
seq_to_character(
  aligned_reads,
  use_CPM,
  alignment_score_cutoff = 0,
  abundance_cutoff = 0
)
}
\arguments{
\item{use_CPM}{Use count per million}

\item{alignment_score_cutoff}{Minimum cutoff for alignment score}

\item{abundance_cutoff}{Minimum cutoff for read count. This parameter are used with use_CPM.}

\item{traceQC_input}{A TraceQC object.}
}
\value{
A data frame that contains the columns:
\itemize{
  \item `type': The type of mutation.
  \item `start': The starting position of mutation event.
  \item `length': The length of mutation.
  \item `mutation_to': A string that shows what mutation is occurred.
  \item `count': The total number of the mutation events from the sample.
}
}
\description{
Identifying mutation events.
}
\examples{
unique_sequences <- group_by(aligned_reads,target_seq,target_ref) \%>\%
summarise(score=max(score)) \%>\%
  ungroup
alignment_score_cutoff <- predict(model,0.4)
mutation <- seq_to_character(unique_sequences,use_CPM=FALSE,
                             alignment_score_cutoff=alignment_score_cutoff)

}
