% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_alignment.R
\name{sequence_alignment}
\alias{sequence_alignment}
\title{Function for a sequence alignment between the reference file and sample.}
\usage{
sequence_alignment(
  input_file,
  ref_file,
  output_file = "aligned_reads.txt",
  match = 2,
  mismatch = -2,
  gapopen = -6,
  gapextension = -0.1,
  return_df = FALSE
)
}
\arguments{
\item{input_file}{A FASTQ file path}

\item{ref_file}{A path of a reference sequence file.}

\item{output_file}{The output path. An output of the alignment will be
stored at the path.}

\item{match}{The score for a correct basepair matching.}

\item{mismatch}{The penalty score for a basepair mismatching.}

\item{gapopen}{The gap opening score for the alignment.}

\item{gapextension}{The gap extension score for the alignment.}

\item{return_df}{A logical argument to report what type of output will
be created from the function.}
}
\value{
It returns a data frame of the alignment resuilt
if `return_df' is `T' and `NULL' otherwise.
}
\description{
The function is an wrapper of a python function which performs
a global pairwise sequence alignment by biopython package.
}
\examples{
\dontrun{
library(TraceQC)
input_file <- system.file("extdata", "test_data",
                          "fastq", "example_small.fastq", package="TraceQC")
ref_file <- system.file("extdata", "test_data", "ref",
                        "ref.txt", package="TraceQC")
output_file <- tempfile()
sequence_alignment(input_file=input_file,
                   ref_file=ref_file,
                   output_file=output_file,
                   return_df=TRUE)
}
}
