% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_alignment.R
\name{sequence_alignment_for_10x}
\alias{sequence_alignment_for_10x}
\title{Function for a sequence alignment between the reference file and sample for 10x data.}
\usage{
sequence_alignment_for_10x(
  input_file,
  ref_file,
  output_file = "aligned_reads.txt",
  match = 2,
  mismatch = -2,
  gapopen = -6,
  gapextension = -0.1,
  ncores = 4,
  return_df = FALSE
)
}
\arguments{
\item{input_file}{A file path of possorted_genome_bam.bam out put by cellranger}

\item{ref_file}{A path of a reference sequence file.}

\item{output_file}{The output path. An output of the alignment will be
stored at the path.}

\item{match}{The score for a correct basepair matching.}

\item{mismatch}{The penalty score for a basepair mismatching.}

\item{gapopen}{The gap opening score for the alignment.}

\item{gapextension}{The gap extension score for the alignment.}

\item{ncores}{The number of cores for the parallel processing}

\item{return_df}{A logical argument to report what type of output will
be created from the function.}
}
\value{
It returns a data frame of the alignment result
if `return_df' is `T' and `NULL' otherwise.
}
\description{
The function is an wrapper of a python function which performs
a global pairwise sequence alignment by biopython package.
}
