% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_alignment.R
\name{sequence_permutation}
\alias{sequence_permutation}
\title{Function for finding threshold of sequence alignment. The function randomly
permutate certain percentage reference sequence and perform global alignment
with the original reference sequence. By use the permutated sequence alignment
score, users can filter the TraceQC alignment result.}
\usage{
sequence_permutation(
  ref_file,
  python_path = "python3",
  match = 2,
  mismatch = -2,
  gapopen = -6,
  gapextension = -0.1,
  penalize_end_gaps = 1,
  read_length = 0,
  permutate_percent = seq(0, 1, length.out = 101),
  n = 2,
  output_file = "alignment_threshold.txt"
)
}
\arguments{
\item{ref_file}{A path of a reference sequence file (required).}

\item{python_path}{The path to Python. (default: "python3").}

\item{match}{The score for a correct basepair matching (default: 2).}

\item{mismatch}{The penalty score for a basepair mismatching (default: -2).}

\item{gapopen}{The gap opening score for the alignment (default: -6).}

\item{gapextension}{The gap extension score for the alignment (default: -0.1).}

\item{penalize_end_gaps}{Whether penalize the end gap, the value should take 0/1 (default: 1).}

\item{output_file}{The output path. An output of the alignment will be
stored at the path (default: "alignment_threshold.txt").}
}
\value{
It returns a data frame of the alignment result with randomly permutated sequence.
}
\description{
Function for finding threshold of sequence alignment. The function randomly
permutate certain percentage reference sequence and perform global alignment
with the original reference sequence. By use the permutated sequence alignment
score, users can filter the TraceQC alignment result.
}
\examples{
ref_file <- system.file("extdata/test_data/ref","ref_carlin.txt",package="TraceQC")
sequence_permutation(ref_file,out="./seq_permutate.txt")
}
