% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_alignment.R
\name{sequnce_alignment_threshold}
\alias{sequnce_alignment_threshold}
\title{Function for finding threshold of sequence alignment. The function randomly
permutate certain percentage reference sequence and perform global alignment
with the original reference sequence. By use the permutated sequence alignment
score, users can filter the TraceQC alignment result.}
\usage{
sequnce_alignment_threshold(
  ref_file,
  match = 2,
  mismatch = -2,
  gapopen = -6,
  gapextension = -0.1,
  permutate_percent = c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6),
  n = 100,
  output_file = "alignment_threshold.txt"
)
}
\arguments{
\item{ref_file}{A path of a reference sequence file.}

\item{match}{The score for a correct basepair matching.}

\item{mismatch}{The penalty score for a basepair mismatching.}

\item{gapopen}{The gap opening score for the alignment.}

\item{gapextension}{The gap extension score for the alignment.}

\item{n}{number of random permutation used for each percentage}

\item{output_file}{The output path. An output dataframe will be
stored at the path.}

\item{corrupted}{percentage The number of cores for the parallel processing}
}
\value{
It returns a data frame of the alignment result
}
\description{
Function for finding threshold of sequence alignment. The function randomly
permutate certain percentage reference sequence and perform global alignment
with the original reference sequence. By use the permutated sequence alignment
score, users can filter the TraceQC alignment result.
}
