% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogetic_tree.R
\name{table_to_phyDat}
\alias{table_to_phyDat}
\title{Convert a binary table of mutation events to a phyDat object}
\usage{
table_to_phyDat(tree_input)
}
\arguments{
\item{tree_input}{A variable which was created by `build_binary_table'}
}
\value{

}
\description{
Convert a binary table of mutation events to a phyDat object
}
