% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraceQC.R
\name{TraceQC}
\alias{TraceQC}
\title{Creating an input object for TraceQC}
\usage{
TraceQC(
  input_file,
  ref_file,
  fastqc_file = NULL,
  alignment_output_file = NULL,
  ncores = 4
)
}
\arguments{
\item{input_file}{A FASTQ file path.}

\item{ref_file}{A path of a reference sequence file.}

\item{fastqc_file}{A path of a FASTQC file.}

\item{alignment_output_file}{A path to store alignment output file.
This will be `NULL' if the output is temporally stored.}
}
\value{
It will return a list from `create_input_object_with_alignment'
}
\description{
Creating an input object for TraceQC
}
\examples{
library(TraceQC)
library(fastqcr)
input_file <- system.file("extdata", "test_data",
                          "fastq", "example.fastq", package="TraceQC")
ref_file <- system.file("extdata", "test_data", "ref",
                        "ref.txt", package="TraceQC")
qc_dir <- tempdir()
fastqc(system.file("extdata", "test_data",
                   "fastq", package = "TraceQC"),
       qc.dir=qc_dir)

input_qc_path <- get_qcpath(input_file, qc_dir)

obj <- TraceQC(input_file = input_file,
               ref_file = ref_file,
               fastqc_file = input_qc_path)

obj$refseq
}
