% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_plots.R
\name{circular_chordgram}
\alias{circular_chordgram}
\title{Display a circos plot with links for a given data frame.}
\usage{
circular_chordgram(df, title, ref, use_log_count = TRUE, count_cutoff = 1)
}
\arguments{
\item{df}{A data frame that contains data to be visualized on the plot}

\item{title}{The main title of the plot}

\item{ref}{An reference object, output of `parse_ref_file`}

\item{use_log_count}{Plot the circular plot useing log()}

\item{count_cutoff}{A cutoff to remove link whose log10-count are less than the value (default: 1).}
}
\value{
It doesn't generate any specific output.
}
\description{
Display a circos plot with links for a given data frame.
}
