% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_plots.R
\name{circular_chordgram}
\alias{circular_chordgram}
\title{Display a circos plot with links for a given data frame.}
\usage{
circular_chordgram(df, title, traceQC_input)
}
\arguments{
\item{df}{a data frame that contains data to be visualized on the plot}

\item{title}{The main title of the plot}

\item{traceQC_input}{A TraceQC object}
}
\value{
It doesn't generate any specific output.
}
\description{
Display a circos plot with links for a given data frame.
}
