% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_plots.R
\name{circular_histogram}
\alias{circular_histogram}
\title{Display a circos plot with a histgoram for a given data frame.}
\usage{
circular_histogram(df, title, ref)
}
\arguments{
\item{df}{a data frame that contains data to be visualized on the plot.}

\item{title}{The main title of the plot.}

\item{ref}{A traceQC object}
}
\value{
It doesn't generate any specific output.
}
\description{
Display a circos plot with a histgoram for a given data frame.
}
