% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_events.R
\name{format_mutation_df}
\alias{format_mutation_df}
\title{Format a mutation data frame for output.}
\usage{
format_mutation_df(mutation_df, is_singlecell)
}
\arguments{
\item{is_singlecell}{If the dataframe is is single-cell.}

\item{mutations}{A data frame of mutations. The output of seq_to_character.}
}
\value{
A formatted data frame of mutations.
}
\description{
Format a mutation data frame for output.
}
\examples{
out_df <- format_mutation_df(mutation,is_singlecell=FALSE)

}
