% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_alignment.R
\name{parse_ref_file}
\alias{parse_ref_file}
\title{Parsing reference sequence file}
\usage{
parse_ref_file(ref_file)
}
\arguments{
\item{ref_file}{A path of a reference sequence file.}
}
\value{
A list with those four elements.
\itemize{
  \item `refseq': The reference sequence.
  \item `regions': Detailed information about the reference sequence.
}
}
\description{
Parsing reference sequence file
}
\examples{
ref_file <- system.file("extdata/test_data/ref","ref_carlin.txt",package="TraceQC")
ref <- parse_ref_file(ref_file)

}
