% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot_construct}
\alias{plot_construct}
\title{Visualization of the construct (reference sequence) information.}
\usage{
plot_construct(traceQC_input)
}
\arguments{
\item{traceQC_input}{an TraceQC object}
}
\value{
it returns A ggplot2 object that shows the construct information.
}
\description{
Visualization of the construct (reference sequence) information.
}
\examples{
library(TraceQC)
data(example_obj)
plot_construct(example_obj)

}
