% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular.R
\name{plot_deletion_hotspot}
\alias{plot_deletion_hotspot}
\title{Display a circos plot that shows overall deletion pattern across the barcodes.}
\usage{
plot_deletion_hotspot(traceQC_input)
}
\arguments{
\item{traceQC_input}{}
}
\value{

}
\description{
Display a circos plot that shows overall deletion pattern across the barcodes.
}
\examples{
data(example_obj)
plot_deletion_hotspot(example_obj)

}
