% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular.R
\name{plot_insertion_hotspot}
\alias{plot_insertion_hotspot}
\title{Display a circos plot that shows overall insertion pattern across the barcodes.}
\usage{
plot_insertion_hotspot(traceQC_input)
}
\arguments{
\item{traceQC_input}{}
}
\value{
It won't return any specific object.
}
\description{
Display a circos plot that shows overall insertion pattern across the barcodes.
}
\examples{
data(example_obj)
plot_insertion_hotspot(example_obj)

}
