% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot_lorenz_curve}
\alias{plot_lorenz_curve}
\title{Drawing Lorenz Curve}
\usage{
plot_lorenz_curve(aligned_reads)
}
\arguments{
\item{aligned_reads}{A aligned_reads dataframe.}
}
\value{
A ggplot2 object that shows Lorenz Curve
}
\description{
The Lorenz curve shows an inequality of barcode distribution of the sample.
}
\examples{
data(example_obj)
plot_lorenz_curve(example_obj)

}
