% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot_score_distribution}
\alias{plot_score_distribution}
\title{Drawing a score distribution plot}
\usage{
plot_score_distribution(aligned_reads)
}
\arguments{
\item{aligned_reads}{A aligned_reads dataframe.}
}
\value{
A ggplot2 object that shows alignment score distribution.
}
\description{
Drawing a score distribution plot
}
\examples{
plot_score_distribution(aligned_reads)

}
